Add 'user authenticate' command
Add the 'user authenticate' command which uses HTTP BASIC authentication with
the users username & password to automatically create client configuration
files for each organization the user is a member of.
Only try delete the $HOME/.cyclid/config symlink if it exists.
Fix the "url" configuration parsing & management.

+++ lib/cyclid/cli/organization.rb
@@ -115,7 +115,7 @@ module Cyclid
           # The configuration file exists and appears to be sane, so switch the
           # 'config' symlink to point to it.
           Dir.chdir(CYCLID_CONFIG_DIR) do
-            File.delete('config')
+            File.delete('config') if File.exist?('config')
             File.symlink(name, 'config')
           end
         end
